Fix #107: Isolate business logic of Data Adapter into separate service - reimplemented with a better interface for consumers
