+++ powerauth-data-adapter/src/main/java/io/getlime/security/powerauth/lib/dataadapter/controller/SMSAuthorizationController.java
@@ -19,11 +19,17 @@ import io.getlime.core.rest.model.base.request.ObjectRequest;
 import io.getlime.core.rest.model.base.response.ObjectResponse;
 import io.getlime.security.powerauth.lib.dataadapter.api.DataAdapter;
 import io.getlime.security.powerauth.lib.dataadapter.exception.SMSAuthorizationFailedException;
+import io.getlime.security.powerauth.lib.dataadapter.impl.validation.CreateSMSAuthorizationRequestValidator;
 import io.getlime.security.powerauth.lib.dataadapter.model.request.CreateSMSAuthorizationRequest;
 import io.getlime.security.powerauth.lib.dataadapter.model.request.VerifySMSAuthorizationRequest;
 import io.getlime.security.powerauth.lib.dataadapter.model.response.CreateSMSAuthorizationResponse;
+import io.getlime.security.powerauth.lib.dataadapter.repository.model.entity.SMSAuthorizationEntity;
+import io.getlime.security.powerauth.lib.dataadapter.service.SMSPersistenceService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BeanPropertyBindingResult;
+import org.springframework.validation.ValidationUtils;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -39,10 +45,14 @@ import org.springframework.web.bind.annotation.ResponseBody;
 @RequestMapping("/api/auth/sms")
 public class SMSAuthorizationController {
 
+    private final SMSPersistenceService smsPersistenceService;
+    private final CreateSMSAuthorizationRequestValidator requestValidator;
     private DataAdapter dataAdapter;
 
     @Autowired
-    public SMSAuthorizationController(DataAdapter dataAdapter) {
+    public SMSAuthorizationController(SMSPersistenceService smsPersistenceService, CreateSMSAuthorizationRequestValidator requestValidator, DataAdapter dataAdapter) {
+        this.smsPersistenceService = smsPersistenceService;
+        this.requestValidator = requestValidator;
         this.dataAdapter = dataAdapter;
     }
 
@@ -54,8 +64,21 @@ public class SMSAuthorizationController {
      */
     @RequestMapping(value = "create", method = RequestMethod.POST)
     public @ResponseBody ObjectResponse<CreateSMSAuthorizationResponse> create(@RequestBody ObjectRequest<CreateSMSAuthorizationRequest> request) throws MethodArgumentNotValidException, SMSAuthorizationFailedException {
-        CreateSMSAuthorizationRequest createSMSAuthorizationRequest = request.getRequestObject();
-        CreateSMSAuthorizationResponse response = dataAdapter.createAuthorizationSMS(createSMSAuthorizationRequest);
+        CreateSMSAuthorizationRequest smsRequest = request.getRequestObject();
+        // input validation is handled by CreateSMSAuthorizationRequestValidator
+        // validation is invoked manually because of the generified Request object
+        BeanPropertyBindingResult result = new BeanPropertyBindingResult(request, "createSMSAuthorizationRequest");
+        ValidationUtils.invokeValidator(requestValidator, request, result);
+        if (result.hasErrors()) {
+            // getEnclosingMethod() on new object returns a reference to current method
+            MethodParameter methodParam = new MethodParameter(new Object() {
+            }.getClass().getEnclosingMethod(), 0);
+            throw new MethodArgumentNotValidException(methodParam, result);
+        }
+        SMSAuthorizationEntity smsEntity = smsPersistenceService.createAuthorizationSMS(smsRequest.getUserId(),
+                smsRequest.getOperationId(), smsRequest.getOperationName(), smsRequest.getOperationFormData(), smsRequest.getLang());
+        dataAdapter.sendAuthorizationSMS(smsEntity.getMessageText(), smsEntity.getUserId());
+        CreateSMSAuthorizationResponse response = new CreateSMSAuthorizationResponse(smsEntity.getMessageId());
         return new ObjectResponse<>(response);
     }
 
@@ -68,7 +91,7 @@ public class SMSAuthorizationController {
     @RequestMapping(value = "verify", method = RequestMethod.POST)
     public @ResponseBody ObjectResponse verify(@RequestBody ObjectRequest<VerifySMSAuthorizationRequest> request) throws SMSAuthorizationFailedException {
         VerifySMSAuthorizationRequest verifyRequest = request.getRequestObject();
-        dataAdapter.verifyAuthorizationSMS(verifyRequest);
+        smsPersistenceService.verifyAuthorizationSMS(verifyRequest.getMessageId(), verifyRequest.getAuthorizationCode());
         return new ObjectResponse();
     }
